requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.12.1
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
alembic==1.13.1
